+++ esp-hal-common/src/uart.rs
@@ -20,8 +20,11 @@ const UART_FIFO_SIZE: u16 = 128;
 /// Custom serial error type
 #[derive(Debug)]
 pub enum Error {
+    InvalidArgument,
     #[cfg(feature = "async")]
     ReadBufferFull,
+    #[cfg(feature = "async")]
+    RxFifoOvf,
 }
 
 /// UART configuration
@@ -248,6 +251,7 @@ impl embedded_hal_1::serial::Error for Error {
 /// UART driver
 pub struct Uart<'d, T> {
     uart: PeripheralRef<'d, T>,
+    at_cmd_config: Option<config::AtCmdConfig>,
 }
 
 impl<'d, T> Uart<'d, T>
@@ -267,7 +271,10 @@ where
     {
         crate::into_ref!(uart);
         uart.enable_peripheral(peripheral_clock_control);
-        let mut serial = Uart { uart };
+        let mut serial = Uart {
+            uart,
+            at_cmd_config: None,
+        };
         serial.uart.disable_rx_interrupts();
         serial.uart.disable_tx_interrupts();
 
@@ -297,7 +304,10 @@ where
     ) -> Self {
         crate::into_ref!(uart);
         uart.enable_peripheral(peripheral_clock_control);
-        let mut serial = Uart { uart };
+        let mut serial = Uart {
+            uart,
+            at_cmd_config: None,
+        };
         serial.uart.disable_rx_interrupts();
         serial.uart.disable_tx_interrupts();
 
@@ -353,10 +363,32 @@ where
             .modify(|_, w| w.sclk_en().set_bit());
 
         self.sync_regs();
+        self.at_cmd_config = Some(config);
     }
 
     /// Configures the RX-FIFO threshold
-    pub fn set_rx_fifo_full_threshold(&mut self, threshold: u16) {
+    ///
+    /// # Errors
+    /// `Err(Error::InvalidArgument)` if provided value exceeds maximum value
+    /// for SOC :
+    /// - `esp32` **0x7F**
+    /// - `esp32c6`, `esp32h2` **0xFF**
+    /// - `esp32c3`, `esp32c2`, `esp32s2` **0x1FF**
+    /// - `esp32s3` **0x3FF**
+    pub fn set_rx_fifo_full_threshold(&mut self, threshold: u16) -> Result<(), Error> {
+        #[cfg(esp32)]
+        const MAX_THRHD: u16 = 0x7F;
+        #[cfg(any(esp32c6, esp32h2))]
+        const MAX_THRHD: u16 = 0xFF;
+        #[cfg(any(esp32c3, esp32c2, esp32s2))]
+        const MAX_THRHD: u16 = 0x1FF;
+        #[cfg(esp32s3)]
+        const MAX_THRHD: u16 = 0x3FF;
+
+        if threshold > MAX_THRHD {
+            return Err(Error::InvalidArgument);
+        }
+
         #[cfg(any(esp32, esp32c6, esp32h2))]
         let threshold: u8 = threshold as u8;
 
@@ -364,6 +396,8 @@ where
             .register_block()
             .conf1
             .modify(|_, w| unsafe { w.rxfifo_full_thrhd().bits(threshold) });
+
+        Ok(())
     }
 
     /// Listen for AT-CMD interrupts
@@ -997,7 +1031,7 @@ mod asynch {
     use core::task::Poll;
 
     use cfg_if::cfg_if;
-    use embassy_futures::select::select;
+    use embassy_futures::select::{select, select3, Either, Either3};
     use embassy_sync::waitqueue::AtomicWaker;
     use procmacros::interrupt;
 
@@ -1025,6 +1059,7 @@ mod asynch {
         TxFiFoEmpty,
         RxFifoFull,
         RxCmdCharDetected,
+        RxFifoOvf,
     }
 
     pub(crate) struct UartFuture<'a, T: Instance> {
@@ -1051,6 +1086,10 @@ mod asynch {
                     .register_block()
                     .int_ena
                     .modify(|_, w| w.at_cmd_char_det_int_ena().set_bit()),
+                Event::RxFifoOvf => instance
+                    .register_block()
+                    .int_ena
+                    .modify(|_, w| w.rxfifo_ovf_int_ena().set_bit()),
             }
 
             Self { event, instance }
@@ -1086,6 +1125,13 @@ mod asynch {
                     .read()
                     .at_cmd_char_det_int_ena()
                     .bit_is_clear(),
+                Event::RxFifoOvf => self
+                    .instance
+                    .register_block()
+                    .int_ena
+                    .read()
+                    .rxfifo_ovf_int_ena()
+                    .bit_is_clear(),
             }
         }
     }
@@ -1110,14 +1156,48 @@ mod asynch {
     where
         T: Instance,
     {
+        /// Read async to buffer slice `buf`. Wait for Rx Fifo Full interrupt
+        /// (set by `set_rx_fifo_full_threshold`) and/or Rx AT_CMD character
+        /// interrupt if `set_at_cmd` was called.
+        ///
+        /// # Params
+        /// - `buf` buffer slice to write the bytes into
+        ///
+        /// # Errors
+        /// - `Err(ReadBufferFull)` if provided buffer slice is not enough to
+        ///   copy all avaialble bytes. Increase buffer slice length.
+        /// - `Err(RxFifoOvf)` when MCU Rx Fifo Overflow interrupt is triggered.
+        ///   To avoid this error, call this function more often.
+        /// - `Err(Error::ReadNoConfig)` if neither `set_rx_fifo_full_threshold`
+        ///   or `set_at_cmd` was called
+        ///
+        /// # Ok
+        /// When succesfull, returns the number of bytes written to
+        /// buf
+
         pub async fn read(&mut self, buf: &mut [u8]) -> Result<usize, Error> {
             let mut read_bytes = 0;
 
-            select(
-                UartFuture::new(Event::RxCmdCharDetected, self.inner()),
-                UartFuture::new(Event::RxFifoFull, self.inner()),
-            )
-            .await;
+            if self.at_cmd_config.is_some() {
+                if let Either3::Third(_) = select3(
+                    UartFuture::new(Event::RxCmdCharDetected, self.inner()),
+                    UartFuture::new(Event::RxFifoFull, self.inner()),
+                    UartFuture::new(Event::RxFifoOvf, self.inner()),
+                )
+                .await
+                {
+                    return Err(Error::RxFifoOvf);
+                }
+            } else {
+                if let Either::Second(_) = select(
+                    UartFuture::new(Event::RxFifoFull, self.inner()),
+                    UartFuture::new(Event::RxFifoOvf, self.inner()),
+                )
+                .await
+                {
+                    return Err(Error::RxFifoOvf);
+                }
+            }
 
             while let Ok(byte) = self.read_byte() {
                 if read_bytes < buf.len() {
@@ -1201,6 +1281,13 @@ mod asynch {
         {
             uart.int_clr.write(|w| w.rxfifo_full_int_clr().set_bit());
             uart.int_ena.write(|w| w.rxfifo_full_int_ena().clear_bit());
+            return true;
+        }
+        if int_ena_val.rxfifo_ovf_int_ena().bit_is_set()
+            && int_raw_val.rxfifo_ovf_int_raw().bit_is_set()
+        {
+            uart.int_clr.write(|w| w.rxfifo_ovf_int_clr().set_bit());
+            uart.int_ena.write(|w| w.rxfifo_ovf_int_ena().clear_bit());
             return true;
         }
         false